Closes #13
refactor fetchTrivia done
